# openapi.yaml
openapi: 3.0.3
info:
  title: Coadjust
  version: 0.1.0
  description: >-
    Claims-workforce operating system that allocates each claim decision to an
    adjuster, a model, or a paired review based on measured combined accuracy.
    Holds the three-mode accuracy baselines, the conduct guardrails that suspend
    automation on deterioration, and the readiness and redeployment ledger for
    handlers released from automated queues.
  x-product-slug: ai-ins-coadjust
  x-source: ai-in-insurance/Accenture-AI-Future-Workforce-Insurance-Transcript/
servers:
  - url: https://api.example.com
...
paths:
...
  /v1/workforce/readiness-assessments:
    get:
      operationId: listReadinessAssessments
      tags: [Workforce]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: roleProfileId
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
            enum: [not_ready, developing, ready, certified, lapsed]
      responses:
        "200":
          description: Readiness per handler per paired role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeReadinessAssessment"
    post:
      operationId: assessHandlerReadiness
      tags: [Workforce]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadinessAssessmentCreate"
      responses:
        "201":
          description: Readiness assessed from production decision quality
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeReadinessAssessment"
  /v1/workforce/redeployments:
    get:
      operationId: listRedeployments
      tags: [Workforce]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            type: string
            enum: [proposed, funded, in_training, placed, lapsed]
      responses:
        "200":
          description: Redeployment ledger page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeRedeploymentAssignment"
    post:
      operationId: assignRedeployment
      tags: [Workforce]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeploymentAssignmentCreate"
      responses:
        "201":
          description: Redeployment assigned against released capacity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeRedeploymentAssignment"
        "422":
          description: >-
            Rejected when released capacity has no destination role or the
            capability path has no funded start date.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/workforce/capacity-plans:
    get:
      operationId: getQueueCapacityPlan
      tags: [Workforce]
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
        - name: proposedPolicyId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Projected handler demand per queue under a proposed policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeQueueCapacityPlan"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          example: EUR
...
    ReadinessAssessment:
      type: object
      required: [id, handlerId, roleProfileId, state]
      properties:
        id:
          type: string
        handlerId:
          type: string
        roleProfileId:
          type: string
        state:
          type: string
          enum: [not_ready, developing, ready, certified, lapsed]
        productionDecisionCount:
          type: integer
        productionAccuracy:
          type: number
          format: float
        overrideUpheldRate:
          type: number
          format: float
        capabilityPathId:
          type: string
        assessedAt:
          type: string
          format: date-time
    ReadinessAssessmentCreate:
      type: object
      required: [handlerId, roleProfileId]
      properties:
        handlerId:
          type: string
        roleProfileId:
          type: string
        capabilityPathId:
          type: string
    RedeploymentAssignment:
      type: object
      required: [id, handlerId, status]
      properties:
        id:
          type: string
        handlerId:
          type: string
        releasedFromDecisionTypeId:
          type: string
        destinationRoleProfileId:
          type: string
        destinationQueue:
          type: string
        capabilityPathId:
          type: string
        fundedStartDate:
          type: string
          format: date
        status:
          type: string
          enum: [proposed, funded, in_training, placed, lapsed]
        releasedCapacityHours:
          type: number
          format: float
        reskillingCost:
          $ref: "#/components/schemas/Money"
        placedAt:
          type: string
          format: date-time
    RedeploymentAssignmentCreate:
      type: object
      required: [handlerId, releasedFromDecisionTypeId, destinationRoleProfileId]
      properties:
        handlerId:
          type: string
        releasedFromDecisionTypeId:
          type: string
        destinationRoleProfileId:
          type: string
        destinationQueue:
          type: string
        capabilityPathId:
          type: string
        fundedStartDate:
          type: string
          format: date
        releasedCapacityHours:
          type: number
          format: float
        reskillingCost:
          $ref: "#/components/schemas/Money"
    QueueCapacityPlan:
      type: object
      properties:
        period:
          type: string
        proposedPolicyId:
          type: string
        queues:
          type: array
          items:
            type: object
            properties:
              queue:
                type: string
              projectedVolume:
                type: integer
              requiredHandlers:
                type: number
                format: float
              certifiedHandlers:
                type: integer
              shortfall:
                type: number
                format: float
        releasedCapacityHours:
          type: number
          format: float
        unassignedCapacityHours:
          type: number
          format: float
...
    DataEnvelopeReadinessAssessment:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/ReadinessAssessment"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    DataEnvelopeRedeploymentAssignment:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/RedeploymentAssignment"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    DataEnvelopeQueueCapacityPlan:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/QueueCapacityPlan"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeReadinessAssessment:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ReadinessAssessment"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ListEnvelopeRedeploymentAssignment:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/RedeploymentAssignment"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...